internal/envoy: split xDS caches into their own types
