+++ internal/envoy/virtualhost.go
@@ -1,0 +1,85 @@
+// Copyright © 2017 Heptio
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package envoy
+
+import (
+	"strings"
+
+	v2 "github.com/envoyproxy/go-control-plane/api"
+	"k8s.io/api/extensions/v1beta1"
+)
+
+// VirtualHostCache manage the contents of the gRPC RDS cache.
+type VirtualHostCache struct {
+	virtualHostCache
+	Cond
+}
+
+// ingressBackendToClusterName renders a cluster name from an Ingress and an IngressBackend.
+func ingressBackendToClusterName(i *v1beta1.Ingress, b *v1beta1.IngressBackend) string {
+	return hashname(60, i.ObjectMeta.Namespace, b.ServiceName, b.ServicePort.String())
+}
+
+// pathToRoute converts a HTTPIngressPath to a partial v2.RouteMatch.
+func pathToRouteMatch(p v1beta1.HTTPIngressPath) *v2.RouteMatch {
+	if p.Path == "" {
+		// If the Path is empty, the k8s spec says
+		// "If unspecified, the path defaults to a catch all sending
+		// traffic to the backend."
+		// We map this it a catch all prefix route.
+		return prefixmatch("/") // match all
+	}
+	// TODO(dfc) handle the case where p.Path does not start with "/"
+	if strings.IndexAny(p.Path, `[(*\`) == -1 {
+		// Envoy requires that regex matches match completely, wheres the
+		// HTTPIngressPath.Path regex only requires a partial match. eg,
+		// "/foo" matches "/" according to k8s rules, but does not match
+		// according to Envoy.
+		// To deal with this we handle the simple case, a Path without regex
+		// characters as a Envoy prefix route.
+		return prefixmatch(p.Path)
+	}
+	// At this point the path is a regex, which we hope is the same between k8s
+	// IEEE 1003.1 POSIX regex, and Envoys Javascript regex.
+	return regexmatch(p.Path)
+}
+
+// prefixmatch returns a RouteMatch for the supplied prefix.
+func prefixmatch(prefix string) *v2.RouteMatch {
+	return &v2.RouteMatch{
+		PathSpecifier: &v2.RouteMatch_Prefix{
+			Prefix: prefix,
+		},
+	}
+}
+
+// regexmatch returns a RouteMatch for the supplied regex.
+func regexmatch(regex string) *v2.RouteMatch {
+	return &v2.RouteMatch{
+		PathSpecifier: &v2.RouteMatch_Regex{
+			Regex: regex,
+		},
+	}
+}
+
+// clusteraction returns a Route_Route action for the supplied cluster.
+func clusteraction(cluster string) *v2.Route_Route {
+	return &v2.Route_Route{
+		Route: &v2.RouteAction{
+			ClusterSpecifier: &v2.RouteAction_Cluster{
+				Cluster: cluster,
+			},
+		},
+	}
+}
